add combine_qa_datasets; add xquad
